Memory optimizations

+++ core/ui/ui.hpp
@@ -1,11 +1,13 @@
 #ifndef _UI_HPP_
 #define _UI_HPP_
 
+#include <stdint.h>
+
 class grid;
 
 void ui ( void );
-void play ( grid G , int size , int alignWinSize , int alignWinTotal );
-void options ( int &defaultheight , int &defaultWidth , int &defaultAlignWinSize , int &defaultAlignWinTotal );
+void play ( grid G , uint8_t size , uint8_t alignWinSize , uint8_t alignWinTotal );
+void options ( uint8_t &defaultheight , uint8_t &defaultWidth , uint8_t &defaultAlignWinSize , uint8_t &defaultAlignWinTotal );
 bool warnExit ( void );
 
 #endif
